# Merge medium and slow scenechange methods into one method