This changeset introduces a significant number of changes to the way that scene detection works internally.

- All frames are now analyzed. Previously, frames that were within `min-keyint` frames from a keyframe were not analyzed at all. Analyzing these frames, even if we know they will not be set as keyframes, improves the accuracy of both the histogram-based cost adjustments as well as flash detection.
- There are several fixes to how flash detection works, as it contained a number of bugs in the previous iteration.
- The histogram-based cost adjustments were improved to give more accurate results, especially in high detail scenes with motion. The threshold bias was adjusted to account for this.
- The last few frames of the video are now ignored, similar to a scene flash. This removes the issue where it was common for the last frame of the video to be set as a keyframe.
- The cost analysis is performed using a hybrid of the previous two methods. The previous "medium" algorithm, as the primary algorithm, will be used for flash detection and will have the cost adjustments applied to it. This was selected as the primary because the intra/inter-cost based algorithm was more accurate at detecting a number of scenarios such as pans, fades, and blur-ins. The previous "slow" algorithm will be used as a secondary check to filter out some false positives that were present in the "medium" algorithm.

Note that the fast algorithm is unchanged and mostly unaffected by this changeset. The only change affecting it is disallowing keyframes at the end of the video. [AWCY](https://beta.arewecompressedyet.com/?job=fast_sc_master%402021-09-20T14%3A32%3A39.865Z&job=fast_sc_sanity_run2%402021-09-21T01%3A44%3A38.657Z)

This changeset was primarily tested against the vimeo-corpus-10s set from AWCY, both within AWCY and manually. The manual testing involved checking each clip to verify that the selected keyframes matched what would be expected by a typical human viewer. For edge cases where it was not clear if a frame should be chosen, BD Rate was considered--if a keyframe improved BD Rate, it should be chosen, otherwise it should not be (for example, frame 31 in slideshow.y4m should be chosen, and frame 40 in ffmpeg.y4m should not be chosen). Manual verification confirmed perfect matches to the expected keyframes on all clips except for brothers.y4m, on which it was impossible to determine expected keyframe placement due to the nature of the video.

AWCY runs can be found here:
- Previous "slow" to new hybrid: [Link](https://beta.arewecompressedyet.com/?job=sc_baseline%402021-09-18T08%3A00%3A40.471Z&job=sc_testing_hybrid%402021-09-19T22%3A59%3A34.162Z)
- Previous "medium" to new hybrid: [Link](https://beta.arewecompressedyet.com/?job=sc_baseline_medium%402021-09-18T08%3A07%3A35.324Z&job=sc_testing_hybrid%402021-09-19T22%3A59%3A34.162Z)

Performance was tested against the 720p full video of Big Buck Bunny, showing the following:
- Fast: 650 fps
- Old Medium: 163 fps
- Old Slow: 272 fps (yes, it wasn't actually slower than medium, it turns out)
- New Method: 151 fps

This equates to a 7.5% drop compared to the previous medium method. This should be considered still fast enough to be useful at most speed levels, and as such, this changeset uses it as the default for speeds 0-9.